+++ src/agens_novel/state/__init__.py
@@ -1,1 +1,0 @@
-"""State package: TypedDict schemas and reducers."""